Update strings.xml
line 676: since this pops up after earning coins, "get" (present tense) doesn't make sense, and "coins earned" is a better representation of what's happening at the time.

lines 1854-1855: line 1854 ("Deadline at") is ungrammatical when the deadline is a specific time (e.g., "deadline at tomorrow"), so I changed it to "Due:". I changed the line to "Due:" as well, because I think it's more succinct (takes up less space) and has less chance of being ungrammatical.

lines 1861-1862: similar to above, "Starts at tomorrow" is ungrammatical, but removing the "on/at" also makes it shorter.

line 363-364: this isn't a translation issue, just a suggestion: "You have a task to complete:" takes up a lot of space in the notification (roughly 2/3 of the notification text on my phone), so a task with medium or long title gest cut off, and the notification cannot be expanded. I think it would be much more efficient for the notification title to be the name of the task, and the description to be the deadline (if there is one; I think this is the most important information to include in a reminder notification). regardless of what you include, the current notification title "LifeUp" and the sentence "You have a task to complete:" doesn't really add any useful information and takes up unnecessary space that could be put to better use.